+++ benches/benchmark.rs
@@ -1,4 +1,5 @@
 use criterion::{black_box, criterion_group, criterion_main};
+use http_range_header::RangeUnsatisfiableError;
 
 pub fn bench(c: &mut criterion::Criterion) {
     c.bench_function("Standard range", |b| {
@@ -27,6 +28,14 @@ pub fn bench(c: &mut criterion::Criterion) {
             http_range_header::parse_range_header(black_box("bytes=0-"))
                 .unwrap()
                 .validate(black_box(10_000))
+        })
+    });
+    c.bench_function("Bad multipart range", |b| {
+        b.iter(|| {
+            assert_eq!(
+                Err(RangeUnsatisfiableError::ZeroSuffix),
+                http_range_header::parse_range_header(black_box("bytes=0-19, -0"))
+            );
         })
     });
 }